% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvExtension-package.R
\docType{package}
\name{flexsurvExtension-package}
\alias{flexsurvExtension}
\alias{flexsurvExtension-package}
\title{flexsurvExtension: Calculate Residual Life Values for Different Distributions}
\description{
A function for calculating mean, median, and/or percentile residual life using the outputs of the flexsurv package. Input the 'flexsurvreg' object, the given 'life' value, the percentile, and the type of residual life, and the function will return your desired values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/an-crawford/flexsurvExtension}
}

}
\author{
\strong{Maintainer}: Andrew Crawford \email{atcrawford18@gmail.com}

}
\keyword{internal}
