% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residLife.R
\name{residLife}
\alias{residLife}
\title{Calculate Residual Life Values}
\usage{
residLife(data, life, p = 0.5, type = "mean")
}
\arguments{
\item{data}{Name of a 'flexsurvreg' object from which data is extracted}

\item{life}{Value of the 'given' used to calculate residual life}

\item{p}{percentile, default is .5}

\item{type}{can be 'mean', 'median', 'percentile', or 'all'. Default is 'mean'}
}
\value{
Residual life values
}
\description{
Calculate Residual Life Values
}
