% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reslife-package.R
\docType{package}
\name{reslife-package}
\alias{reslife}
\alias{reslife-package}
\title{reslife: Calculate Mean Residual Life (MRL) and Related Values for Different Distributions}
\description{
A pair of functions for calculating mean residual life (MRL), median residual life, and percentile residual life using the outputs of either the 'flexsurv' package or parameters provided by the user. Input information about the distribution, the given 'life' value, the percentile, and the type of residual life, and the function will return your desired values. For the 'flexsurv' option, the function allows the user to input their own data for making predictions. This function is based on Jackson (2016) \doi{10.18637/jss.v070.i08}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/an-crawford/reslife}
}

}
\author{
\strong{Maintainer}: Andrew Crawford \email{atcrawford18@gmail.com}

Authors:
\itemize{
  \item Zekai Wang \email{wangzekai1203@gmail.com}
  \item Ka Lok Lee \email{kaloklee@gmail.com}
}

}
\keyword{internal}
