% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residLife.R
\name{residLife}
\alias{residLife}
\title{Calculate Residual Life Values}
\usage{
residLife(data, life, p = 0.5, type = "mean", newdata = data.frame())
}
\arguments{
\item{data}{Name of a 'flexsurvreg' object from which data is extracted}

\item{life}{Value of the 'given' used to calculate residual life}

\item{p}{percentile, default is .5}

\item{type}{can be 'mean', 'median', 'percentile', or 'all'. Default is
'mean'}

\item{newdata}{a data frame containing new data values to calculate residual
life for. Default is a blank data frame.}
}
\value{
Residual life values
}
\description{
Calculate Residual Life Values
}
\examples{
residLife(flexsurvreg, 6, .75, 'all')
residLife(flexsurvreg, 3, type = 'median')
}
