% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residlife.R
\name{residlife}
\alias{residlife}
\title{Calculating Residual Life Values}
\usage{
residlife(min, max, distribution, parameters, p = 0.5, type = "mean")
}
\arguments{
\item{min}{Minimum value for which you want to see the MRL for}

\item{max}{Maximum value for which you want to see the MRL for}

\item{distribution}{Name of the distribution. Needs to be one of "weibull", "gompertz",
"gamma", "gengamma.orig", "exponential", "lnorm", or "llogis".}

\item{parameters}{Parameters of the survival function. Needs to be inputted
in order as a vector, with the name of the parameter included.}

\item{p}{Percentile at which to calculate residual life. Default is .5}

\item{type}{Type of residual life outputted. Must be "mean", "median", "percentile",
or "all". Default is "mean".}
}
\value{
The residual life for a specified sequence of values
}
\description{
Calculating Residual Life Values
}
\examples{
residlife(0, 60, 'weibull', c(shape = 1.2, scale = 3))
residlife(15, 35, 'gamma', c(shape = 1.2, rate =  1.7), p = .25, type ='all')
}
